LPS-58771 Move creation to SoyManager
