+++ modules/apps/foundation/portal-template/portal-template-soy/src/main/java/com/liferay/portal/template/soy/internal/SoyManager.java
@@ -28,8 +28,14 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 
+import org.osgi.framework.Bundle;
+import org.osgi.framework.BundleContext;
+import org.osgi.framework.wiring.BundleCapability;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
+import org.osgi.util.tracker.BundleTracker;
 
 /**
  * @author Bruno Basto
@@ -77,6 +83,22 @@ public class SoyManager extends BaseMultiTemplateManager {
 		super.setTemplateContextHelper(templateContextHelper);
 	}
 
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		int stateMask = Bundle.ACTIVE | Bundle.RESOLVED;
+
+		_bundleTracker = new BundleTracker<>(
+			bundleContext, stateMask,
+			new SoyCapabilityBundleTrackerCustomizer(_portalCache));
+
+		_bundleTracker.open();
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		_bundleTracker.close();
+	}
+
 	@Override
 	protected Template doGetTemplate(
 		List<TemplateResource> templateResources,
@@ -96,6 +118,7 @@ public class SoyManager extends BaseMultiTemplateManager {
 		return template;
 	}
 
+	private BundleTracker<List<BundleCapability>> _bundleTracker;
 	private PortalCache<HashSet<TemplateResource>, SoyTofuCacheBag>
 		_portalCache;
